+++ src/main/java/com/modsen/pizzeria/aspect/CheckEmailMatchAspect.java
@@ -1,9 +1,9 @@
-package com.modsen.pizzeria.aspects;
+package com.modsen.pizzeria.aspect;
 
-import com.modsen.pizzeria.domain.Order;
+import com.modsen.pizzeria.domain.User;
 import com.modsen.pizzeria.exception.AccessDeniedException;
 import com.modsen.pizzeria.exception.ResourceNotFoundException;
-import com.modsen.pizzeria.repository.OrderRepository;
+import com.modsen.pizzeria.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
@@ -21,16 +21,16 @@ import static com.modsen.pizzeria.error.ErrorMessages.RESOURCE_NOT_FOUND_MESSAGE
 @Aspect
 @Component
 @RequiredArgsConstructor
-public class CheckOrderOwnerAspect {
+public class CheckEmailMatchAspect {
 
-    private final OrderRepository orderRepository;
+    private final UserRepository userRepository;
 
-    @Pointcut("@annotation(com.modsen.pizzeria.aspects.annotations.OrderOwnerOrAdminAccess)")
-    public void checkOrderOwner() {
+    @Pointcut("@annotation(com.modsen.pizzeria.aspect.annotation.EmailMatchOrAdminAccess)")
+    public void checkEmailMatch() {
     }
 
-    @Before("checkOrderOwner() && args(orderId,..)")
-    public void before(JoinPoint joinPoint, Long orderId) {
+    @Before("checkEmailMatch() && args(userId,..)")
+    public void before(JoinPoint joinPoint, Long userId) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String currentUserEmail = authentication.getName();
 
@@ -38,12 +38,12 @@ public class CheckOrderOwnerAspect {
                 .map(GrantedAuthority::getAuthority)
                 .toList();
 
-        Order order = orderRepository.findById(orderId)
+        User userToOperate = userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException(
-                        String.format(RESOURCE_NOT_FOUND_MESSAGE, "Order", orderId)
+                        String.format(RESOURCE_NOT_FOUND_MESSAGE, "User", userId)
                 ));
 
-        if (!currentUserEmail.equals(order.getUser().getEmail())
+        if (!currentUserEmail.equals(userToOperate.getEmail())
                 && !userAuthorities.contains("ROLE_" + ADMIN.name())) {
             throw new AccessDeniedException(ACCESS_DENIED_ERROR_MESSAGE);
         }
