+++ src/main/java/com/modsen/pizzeria/aspect/CheckOrderOwnerAspect.java
@@ -1,9 +1,9 @@
-package com.modsen.pizzeria.aspects;
+package com.modsen.pizzeria.aspect;
 
-import com.modsen.pizzeria.domain.User;
+import com.modsen.pizzeria.domain.Order;
 import com.modsen.pizzeria.exception.AccessDeniedException;
 import com.modsen.pizzeria.exception.ResourceNotFoundException;
-import com.modsen.pizzeria.repository.UserRepository;
+import com.modsen.pizzeria.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
@@ -21,16 +21,16 @@ import static com.modsen.pizzeria.error.ErrorMessages.RESOURCE_NOT_FOUND_MESSAGE
 @Aspect
 @Component
 @RequiredArgsConstructor
-public class CheckEmailMatchAspect {
+public class CheckOrderOwnerAspect {
 
-    private final UserRepository userRepository;
+    private final OrderRepository orderRepository;
 
-    @Pointcut("@annotation(com.modsen.pizzeria.aspects.annotations.EmailMatchOrAdminAccess)")
-    public void checkEmailMatch() {
+    @Pointcut("@annotation(com.modsen.pizzeria.aspect.annotation.OrderOwnerOrAdminAccess)")
+    public void checkOrderOwner() {
     }
 
-    @Before("checkEmailMatch() && args(userId,..)")
-    public void before(JoinPoint joinPoint, Long userId) {
+    @Before("checkOrderOwner() && args(orderId,..)")
+    public void before(JoinPoint joinPoint, Long orderId) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String currentUserEmail = authentication.getName();
 
@@ -38,12 +38,12 @@ public class CheckEmailMatchAspect {
                 .map(GrantedAuthority::getAuthority)
                 .toList();
 
-        User userToOperate = userRepository.findById(userId)
+        Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new ResourceNotFoundException(
-                        String.format(RESOURCE_NOT_FOUND_MESSAGE, "User", userId)
+                        String.format(RESOURCE_NOT_FOUND_MESSAGE, "Order", orderId)
                 ));
 
-        if (!currentUserEmail.equals(userToOperate.getEmail())
+        if (!currentUserEmail.equals(order.getUser().getEmail())
                 && !userAuthorities.contains("ROLE_" + ADMIN.name())) {
             throw new AccessDeniedException(ACCESS_DENIED_ERROR_MESSAGE);
         }
