+++ src/main/java/com/modsen/pizzeria/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.modsen.pizzeria.controllers;
+package com.modsen.pizzeria.controller;
 
 import com.modsen.pizzeria.swagger.UserApi;
 import com.modsen.pizzeria.dto.request.UpdateUserRequest;
