--- src/main/java/com/modsen/pizzeria/mappers/OrderItemMapper.java
@@ -1,15 +1,0 @@
-package com.modsen.pizzeria.mappers;
-
-import com.modsen.pizzeria.domain.OrderItem;
-import com.modsen.pizzeria.dto.response.OrderItemResponse;
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface OrderItemMapper {
-
-    OrderItemResponse toOrderItemDto(OrderItem orderItem);
-
-    OrderItem toOrderItem(OrderItemResponse orderItemResponse);
-
-}
